Move network functions to separate file.
